Restore the macOS build: juce::jmax<size_t> instantiates JUCE's SIMD overload
The macOS "Build & Validate" job has failed to COMPILE tests/state_tests.cpp on
every push since f50d6c2, while Linux and Windows stayed green. One line did it:

    return std::sqrt (s / juce::jmax<size_t> (1, v.size()));

juce::jmax is not one function. juce_dsp declares a second overload taking
dsp::SIMDRegister<Type> (juce_SIMDRegister_Impl.h:185). An EXPLICIT template
argument is substituted into every candidate, so the compiler must form
SIMDRegister<size_t> — and completing that class needs SIMDNativeOps<size_t>,
which exists only when size_t happens to name one of the ten types JUCE
specialises. A parameter of incomplete class type must be completed, and that
completion is outside the immediate context, so it is a hard error rather than a
quiet SFINAE removal.

    Linux  (LP64, glibc):  uint64_t IS unsigned long IS size_t -> complete
    macOS  (LP64, libc++): uint64_t is unsigned long long,
                           size_t is unsigned long             -> INCOMPLETE

So the defect is invisible to a Linux compiler no matter WHICH compiler it is:
the typedef diverges, not the front end. A Linux+Clang job would not have caught
this, and saying so matters more than adding one.

Verified by compilation rather than inference. On Linux, `long long` is the exact
structural analogue of macOS's `unsigned long` — a 64-bit integer type that is
not what <cstdint> spells int64_t there. juce::jmax<long long>(1, 7) reproduces
the identical incomplete-type error here, and juce::jmax((long long) 1, 7LL)
compiles clean: deducing Type from a scalar against parameter SIMDRegister<Type>
FAILS, which removes the SIMD candidate before the class is ever completed.

THE FIX is that deduced form, at the one broken site and at the ten sites that
were merely near-misses. Those ten passed macOS for a principled reason, not by
luck — they spell int64_t, and JUCE spells the specialisation SIMDNativeOps<int64_t>,
so it tracks whatever the platform typedef denotes. They change anyway because
the FORM is what admits the mistake, and every one is bit-identical for scalars.

THE GUARD is scripts/check-portability.py, and it is a lint on purpose: no build
job on this machine can see this class. It rejects an explicit template argument
on the closed set {jmin, jmax, snapToZero} — the juce names juce_dsp also
overloads for SIMDRegister, swept from the JUCE headers rather than assumed;
jlimit, jmap, findMinimum, approximatelyEqual and roundToInt have no such
overload and are deliberately NOT listed, because a lint that flags safe code
gets switched off. Comments and string literals are blanked before matching, so
the rule does not fire on the prose explaining it. Mutation-verified: reinstating
the exact 5694 line fails the lint at that line and nothing else.

--compile-canary proves the premise instead of trusting the docstring: it
compiles the deduced and explicit forms against the pinned JUCE and requires the
first to succeed and the second to fail. A JUCE bump that changes the overload
set fails the canary rather than silently voiding the lint.

Also removes two (size_t) casts on juce::dsp::AudioBlock::getSample/setSample
(AnabasisEngine.cpp:742, :837), which take int — the only two real warnings the
macOS log carried, and they sit in the per-channel region access KI-009 points
at. Bit-identical; the casts narrowed straight back.

Suites 293 + 722 green, and both are now VALGRIND-CLEAN under memcheck with
--track-origins (0 errors from 0 contexts each), which is the first direct
evidence against the read-before-write hypothesis for KI-009 on every path the
suites cover.

+++ src/dsp/AnabasisEngine.cpp
@@ -739,7 +739,7 @@ void AnabasisEngine::processChunk (juce::AudioBuffer<float>& buffer, const int s
         float frame[kMaxChannels] = {};
         for (int ch = 0; ch < nCh; ++ch)
         {
-            frame[ch] = region.getSample ((size_t) ch, (size_t) i);
+            frame[ch] = region.getSample (ch, i);
             if (! std::isfinite (frame[ch]))
             {
                 frame[ch] = 0.0f;              // the oversampler's own filters
@@ -834,7 +834,7 @@ void AnabasisEngine::processChunk (juce::AudioBuffer<float>& buffer, const int s
         for (int ch = 0; ch < nCh; ++ch)
         {
             const float delayedWet = wetRing.getSample (ch, readPos);
-            region.setSample ((size_t) ch, (size_t) i,
+            region.setSample (ch, i,
                               juce::exactlyEqual (gains[ch], 1.0f) ? delayedWet
                                                                    : delayedWet * gains[ch]);
         }
